requirements-spec.txt
# Base requirements
django==4.2.2
psycopg[c]==3.1.9
django-environ==0.10.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
markdown-it-py==3.0.0
django-formtools==2.4.1

# Development requirements
coverage==7.2.7
pylint==2.17.4
pylint-django==2.5.3
django-debug-toolbar==4.1.0
